Added Dynamics
Removed window::View,
Made keyevent continuous, and event now toggles,

+++ src/game/hero.h
@@ -17,8 +17,8 @@ public:
   Hero(std::string const& filename, display::Window& window, Scene& scene, Collision& collision, event::Signal& pause);
   Hero(json::JSON const& json, display::Window& window, Scene& scene, Collision& collision, event::Signal& pause);
   Hero(void);
-  int& x(void);
-  int& y(void);
+  float& x(void);
+  float& y(void);
   void End(event::Command const& command);
 
   ~Hero(void);
